removed hook; created structured format func.
