+++ src/views/Transactions/formatTransactionData.tsx
@@ -1,6 +1,6 @@
 import { TableLogo, TableLink } from "./TransactionTable.styles";
 import { shortenTransactionHash } from "utils/format";
-import { IRow } from "components/Table/Table";
+import { ICell, IRow } from "components/Table/Table";
 
 import arbLogo from "assets/arbitrum-logo.svg";
 import umaLogo from "assets/UMA-round.svg";
@@ -8,7 +8,19 @@ import ethLogo from "assets/ethereum-logo.svg";
 
 // Stub of function.
 // Will take View Model Transaction as arg
-export function formatTransactions() {
+export default function formatTransactionsData() {
+  const rows = formatRows();
+  const filledTx = rows.filter((x, i) => i !== 0);
+  const ongoingTx = rows.filter((x, i) => i === 0);
+  return {
+    headers,
+    ongoingTx,
+    filledTx,
+  };
+}
+
+// Will take a TransactionsArg
+function formatRows() {
   const rows: IRow[] = [
     {
       cells: [
@@ -244,5 +256,48 @@ export function formatTransactions() {
     },
   ];
 
-  return { rows };
+  return rows;
 }
+
+const headers: ICell[] = [
+  {
+    size: "lg",
+    value: "Deposit time",
+    cellClassName: "header-cell",
+  },
+  {
+    size: "sm",
+    value: "Status",
+    cellClassName: "header-cell",
+  },
+  {
+    size: "sm",
+    value: "Filled %",
+    cellClassName: "header-cell",
+  },
+  {
+    size: "sm",
+    value: "Source",
+    cellClassName: "header-cell",
+  },
+  {
+    size: "sm",
+    value: "Destination",
+    cellClassName: "header-cell",
+  },
+  {
+    size: "sm",
+    value: "Asset",
+    cellClassName: "header-cell",
+  },
+  {
+    size: "sm",
+    value: "Amount",
+    cellClassName: "header-cell",
+  },
+  {
+    size: "sm",
+    value: "Deposit tx",
+    cellClassName: "header-cell",
+  },
+];
